# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type HuntMode = (typeof HuntMode)[keyof typeof HuntMode];

export const HuntMode = {
  code4rena: "code4rena",
  immunefi: "immunefi",
} as const;

export type HuntStatus = (typeof HuntStatus)[keyof typeof HuntStatus];

export const HuntStatus = {
  pending: "pending",
  running: "running",
  complete: "complete",
  failed: "failed",
} as const;

export type FindingSeverity =
  (typeof FindingSeverity)[keyof typeof FindingSeverity];

export const FindingSeverity = {
  critical: "critical",
  high: "high",
  medium: "medium",
  low: "low",
  informational: "informational",
  gas: "gas",
} as const;

export interface Finding {
  id: string;
  severity: FindingSeverity;
  title: string;
  contract: string;
  function?: string | null;
  description: string;
  impact: string;
  recommendation: string;
  category: string;
  codeSnippet?: string | null;
  proofOfConcept?: string | null;
}

export interface Hunt {
  id: string;
  repoUrl: string;
  repoName: string;
  mode: HuntMode;
  model?: string;
  status: HuntStatus;
  contractsFound?: number | null;
  findings?: Finding[] | null;
  reportMarkdown?: string | null;
  errorMessage?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateHuntBodyMode =
  (typeof CreateHuntBodyMode)[keyof typeof CreateHuntBodyMode];

export const CreateHuntBodyMode = {
  code4rena: "code4rena",
  immunefi: "immunefi",
} as const;

export type CreateHuntBodyModel =
  (typeof CreateHuntBodyModel)[keyof typeof CreateHuntBodyModel];

export const CreateHuntBodyModel = {
  "anthropic/claude-haiku-45": "anthropic/claude-haiku-4.5",
  "anthropic/claude-sonnet-4": "anthropic/claude-sonnet-4",
  "anthropic/claude-opus-4": "anthropic/claude-opus-4",
} as const;

export interface CreateHuntBody {
  repoUrl: string;
  mode: CreateHuntBodyMode;
  model?: CreateHuntBodyModel;
}

export interface HuntStats {
  totalHunts: number;
  completedHunts: number;
  totalFindings: number;
  criticalFindings: number;
  highFindings: number;
  recentHunts: Hunt[];
}

export type DonationStatusTier =
  (typeof DonationStatusTier)[keyof typeof DonationStatusTier];

export const DonationStatusTier = {
  free: "free",
  small: "small",
  medium: "medium",
  lifetime: "lifetime",
} as const;

export type DonationStatusTiersSmall = {
  ethMin: number;
  hunts: number;
  days: number;
};

export type DonationStatusTiersMedium = {
  ethMin: number;
  hunts: number;
  days: number;
};

export type DonationStatusTiersLifetime = {
  ethMin: number;
  hunts: string;
};

export type DonationStatusTiers = {
  small: DonationStatusTiersSmall;
  medium: DonationStatusTiersMedium;
  lifetime: DonationStatusTiersLifetime;
};

export interface DonationStatus {
  tier: DonationStatusTier;
  /** null means unlimited */
  huntsRemaining?: number | null;
  huntsUsed: number;
  expiresAt?: string | null;
  isSponsor: boolean;
  donationAddress: string;
  tiers: DonationStatusTiers;
}

export interface VerifyDonationBody {
  /** Ethereum mainnet transaction hash */
  txHash: string;
}

export interface Sponsor {
  address: string;
  since: string;
}

export interface ApiError {
  error: string;
}