# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: hunts
    description: Bloodhound security audit hunts
  - name: donations
    description: ETH donation verification and hunt quota
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /hunts:
    get:
      operationId: listHunts
      tags: [hunts]
      summary: List all hunts
      responses:
        "200":
          description: List of hunts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Hunt"
    post:
      operationId: createHunt
      tags: [hunts]
      summary: Start a new security audit hunt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHuntBody"
      responses:
        "201":
          description: Hunt created and queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hunt"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /hunts/{id}:
    get:
      operationId: getHunt
      tags: [hunts]
      summary: Get a hunt by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Hunt details with findings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hunt"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /hunts/{id}/progress:
    get:
      operationId: getHuntProgress
      tags: [hunts]
      summary: SSE stream of hunt progress events
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: SSE stream of progress events
          content:
            text/event-stream: {}

  /donations/status:
    get:
      operationId: getDonationStatus
      tags: [donations]
      summary: Get hunt quota and donation tier for the current IP
      responses:
        "200":
          description: Current donation status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DonationStatus"

  /donations/verify:
    post:
      operationId: verifyDonation
      tags: [donations]
      summary: Verify an ETH donation transaction and unlock hunt quota
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyDonationBody"
      responses:
        "200":
          description: Donation verified and quota updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DonationStatus"
        "400":
          description: Invalid or unrecognized transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /donations/sponsors:
    get:
      operationId: listSponsors
      tags: [donations]
      summary: List lifetime sponsors
      responses:
        "200":
          description: List of sponsor addresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sponsor"

  /hunts/stats:
    get:
      operationId: getHuntStats
      tags: [hunts]
      summary: Get aggregate statistics across all hunts
      responses:
        "200":
          description: Aggregate stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HuntStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Hunt:
      type: object
      properties:
        id:
          type: string
        repoUrl:
          type: string
        repoName:
          type: string
        mode:
          type: string
          enum: [code4rena, immunefi]
        model:
          type: string
          default: anthropic/claude-sonnet-4
        status:
          type: string
          enum: [pending, running, complete, failed]
        contractsFound:
          type: integer
          nullable: true
        findings:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/Finding"
        reportMarkdown:
          type: string
          nullable: true
        errorMessage:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - repoUrl
        - repoName
        - mode
        - status
        - createdAt
        - updatedAt

    Finding:
      type: object
      properties:
        id:
          type: string
        severity:
          type: string
          enum: [critical, high, medium, low, informational, gas]
        title:
          type: string
        contract:
          type: string
        function:
          type: string
          nullable: true
        description:
          type: string
        impact:
          type: string
        recommendation:
          type: string
        category:
          type: string
        codeSnippet:
          type: string
          nullable: true
        proofOfConcept:
          type: string
          nullable: true
      required:
        - id
        - severity
        - title
        - contract
        - description
        - impact
        - recommendation
        - category

    CreateHuntBody:
      type: object
      properties:
        repoUrl:
          type: string
        mode:
          type: string
          enum: [code4rena, immunefi]
        model:
          type: string
          enum:
            - anthropic/claude-haiku-4.5
            - anthropic/claude-sonnet-4
            - anthropic/claude-opus-4
          default: anthropic/claude-sonnet-4
      required:
        - repoUrl
        - mode

    HuntStats:
      type: object
      properties:
        totalHunts:
          type: integer
        completedHunts:
          type: integer
        totalFindings:
          type: integer
        criticalFindings:
          type: integer
        highFindings:
          type: integer
        recentHunts:
          type: array
          items:
            $ref: "#/components/schemas/Hunt"
      required:
        - totalHunts
        - completedHunts
        - totalFindings
        - criticalFindings
        - highFindings
        - recentHunts

    DonationStatus:
      type: object
      properties:
        tier:
          type: string
          enum: [free, small, medium, lifetime]
        huntsRemaining:
          type: integer
          nullable: true
          description: null means unlimited
        huntsUsed:
          type: integer
        expiresAt:
          type: string
          format: date-time
          nullable: true
        isSponsor:
          type: boolean
        donationAddress:
          type: string
        tiers:
          type: object
          properties:
            small:
              type: object
              properties:
                ethMin: { type: number }
                hunts: { type: integer }
                days: { type: integer }
              required: [ethMin, hunts, days]
            medium:
              type: object
              properties:
                ethMin: { type: number }
                hunts: { type: integer }
                days: { type: integer }
              required: [ethMin, hunts, days]
            lifetime:
              type: object
              properties:
                ethMin: { type: number }
                hunts: { type: string }
              required: [ethMin, hunts]
          required: [small, medium, lifetime]
      required:
        - tier
        - huntsUsed
        - isSponsor
        - donationAddress
        - tiers

    VerifyDonationBody:
      type: object
      properties:
        txHash:
          type: string
          description: Ethereum mainnet transaction hash
      required:
        - txHash

    Sponsor:
      type: object
      properties:
        address:
          type: string
        since:
          type: string
          format: date-time
      required:
        - address
        - since

    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error