# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all hunts
 */
export const listHuntsResponseModelDefault = `anthropic/claude-sonnet-4`;

export const ListHuntsResponseItem = zod.object({
  id: zod.string(),
  repoUrl: zod.string(),
  repoName: zod.string(),
  mode: zod.enum(["code4rena", "immunefi"]),
  model: zod.string().default(listHuntsResponseModelDefault),
  status: zod.enum(["pending", "running", "complete", "failed"]),
  contractsFound: zod.number().nullish(),
  findings: zod
    .array(
      zod.object({
        id: zod.string(),
        severity: zod.enum([
          "critical",
          "high",
          "medium",
          "low",
          "informational",
          "gas",
        ]),
        title: zod.string(),
        contract: zod.string(),
        function: zod.string().nullish(),
        description: zod.string(),
        impact: zod.string(),
        recommendation: zod.string(),
        category: zod.string(),
        codeSnippet: zod.string().nullish(),
        proofOfConcept: zod.string().nullish(),
      }),
    )
    .nullish(),
  reportMarkdown: zod.string().nullish(),
  errorMessage: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListHuntsResponse = zod.array(ListHuntsResponseItem);

/**
 * @summary Start a new security audit hunt
 */
export const createHuntBodyModelDefault = `anthropic/claude-sonnet-4`;

export const CreateHuntBody = zod.object({
  repoUrl: zod.string(),
  mode: zod.enum(["code4rena", "immunefi"]),
  model: zod
    .enum([
      "anthropic/claude-haiku-4.5",
      "anthropic/claude-sonnet-4",
      "anthropic/claude-opus-4",
    ])
    .default(createHuntBodyModelDefault),
});

/**
 * @summary Get a hunt by ID
 */
export const GetHuntParams = zod.object({
  id: zod.coerce.string(),
});

export const getHuntResponseModelDefault = `anthropic/claude-sonnet-4`;

export const GetHuntResponse = zod.object({
  id: zod.string(),
  repoUrl: zod.string(),
  repoName: zod.string(),
  mode: zod.enum(["code4rena", "immunefi"]),
  model: zod.string().default(getHuntResponseModelDefault),
  status: zod.enum(["pending", "running", "complete", "failed"]),
  contractsFound: zod.number().nullish(),
  findings: zod
    .array(
      zod.object({
        id: zod.string(),
        severity: zod.enum([
          "critical",
          "high",
          "medium",
          "low",
          "informational",
          "gas",
        ]),
        title: zod.string(),
        contract: zod.string(),
        function: zod.string().nullish(),
        description: zod.string(),
        impact: zod.string(),
        recommendation: zod.string(),
        category: zod.string(),
        codeSnippet: zod.string().nullish(),
        proofOfConcept: zod.string().nullish(),
      }),
    )
    .nullish(),
  reportMarkdown: zod.string().nullish(),
  errorMessage: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary SSE stream of hunt progress events
 */
export const GetHuntProgressParams = zod.object({
  id: zod.coerce.string(),
});

/**
 * @summary Get hunt quota and donation tier for the current IP
 */
export const GetDonationStatusResponse = zod.object({
  tier: zod.enum(["free", "small", "medium", "lifetime"]),
  huntsRemaining: zod.number().nullish().describe("null means unlimited"),
  huntsUsed: zod.number(),
  expiresAt: zod.coerce.date().nullish(),
  isSponsor: zod.boolean(),
  donationAddress: zod.string(),
  tiers: zod.object({
    small: zod.object({
      ethMin: zod.number(),
      hunts: zod.number(),
      days: zod.number(),
    }),
    medium: zod.object({
      ethMin: zod.number(),
      hunts: zod.number(),
      days: zod.number(),
    }),
    lifetime: zod.object({
      ethMin: zod.number(),
      hunts: zod.string(),
    }),
  }),
});

/**
 * @summary Verify an ETH donation transaction and unlock hunt quota
 */
export const VerifyDonationBody = zod.object({
  txHash: zod.string().describe("Ethereum mainnet transaction hash"),
});

export const VerifyDonationResponse = zod.object({
  tier: zod.enum(["free", "small", "medium", "lifetime"]),
  huntsRemaining: zod.number().nullish().describe("null means unlimited"),
  huntsUsed: zod.number(),
  expiresAt: zod.coerce.date().nullish(),
  isSponsor: zod.boolean(),
  donationAddress: zod.string(),
  tiers: zod.object({
    small: zod.object({
      ethMin: zod.number(),
      hunts: zod.number(),
      days: zod.number(),
    }),
    medium: zod.object({
      ethMin: zod.number(),
      hunts: zod.number(),
      days: zod.number(),
    }),
    lifetime: zod.object({
      ethMin: zod.number(),
      hunts: zod.string(),
    }),
  }),
});

/**
 * @summary List lifetime sponsors
 */
export const ListSponsorsResponseItem = zod.object({
  address: zod.string(),
  since: zod.coerce.date(),
});
export const ListSponsorsResponse = zod.array(ListSponsorsResponseItem);

/**
 * @summary Get aggregate statistics across all hunts
 */
export const getHuntStatsResponseRecentHuntsItemModelDefault = `anthropic/claude-sonnet-4`;

export const GetHuntStatsResponse = zod.object({
  totalHunts: zod.number(),
  completedHunts: zod.number(),
  totalFindings: zod.number(),
  criticalFindings: zod.number(),
  highFindings: zod.number(),
  recentHunts: zod.array(
    zod.object({
      id: zod.string(),
      repoUrl: zod.string(),
      repoName: zod.string(),
      mode: zod.enum(["code4rena", "immunefi"]),
      model: zod
        .string()
        .default(getHuntStatsResponseRecentHuntsItemModelDefault),
      status: zod.enum(["pending", "running", "complete", "failed"]),
      contractsFound: zod.number().nullish(),
      findings: zod
        .array(
          zod.object({
            id: zod.string(),
            severity: zod.enum([
              "critical",
              "high",
              "medium",
              "low",
              "informational",
              "gas",
            ]),
            title: zod.string(),
            contract: zod.string(),
            function: zod.string().nullish(),
            description: zod.string(),
            impact: zod.string(),
            recommendation: zod.string(),
            category: zod.string(),
            codeSnippet: zod.string().nullish(),
            proofOfConcept: zod.string().nullish(),
          }),
        )
        .nullish(),
      reportMarkdown: zod.string().nullish(),
      errorMessage: zod.string().nullish(),
      createdAt: zod.coerce.date(),
      updatedAt: zod.coerce.date(),
    }),
  ),
});